Extract loaders from ipv8_module_catalog.py
