+++ src/tribler-core/tribler_core/modules/ipv8_module_catalog.py
@@ -2,7 +2,7 @@
 import inspect
 import sys
 
-from ipv8.loader import CommunityLauncher, after, kwargs, overlay, precondition, set_in_session, walk_strategy
+from ipv8.loader import CommunityLauncher, kwargs, overlay, precondition, set_in_session, walk_strategy
 from ipv8.peer import Peer
 
 INFINITE = -1
@@ -43,41 +43,6 @@ def dht_discovery_community():
     return DHTDiscoveryCommunity
 
 
-def tribler_tunnel_community():
-    from tribler_core.modules.tunnel.community.community import TriblerTunnelCommunity
-    return TriblerTunnelCommunity
-
-
-def bandwidth_accounting_community():
-    from tribler_core.modules.bandwidth_accounting.community import BandwidthAccountingCommunity
-    return BandwidthAccountingCommunity
-
-
-def bandwidth_accounting_testnet_community():
-    from tribler_core.modules.bandwidth_accounting.community import BandwidthAccountingTestnetCommunity
-    return BandwidthAccountingTestnetCommunity
-
-
-def popularity_community():
-    from tribler_core.modules.popularity.community import PopularityCommunity
-    return PopularityCommunity
-
-
-def giga_channel_community():
-    from tribler_core.modules.metadata_store.community.gigachannel_community import GigaChannelCommunity
-    return GigaChannelCommunity
-
-
-def giga_channel_testnet_community():
-    from tribler_core.modules.metadata_store.community.gigachannel_community import GigaChannelTestnetCommunity
-    return GigaChannelTestnetCommunity
-
-
-def tribler_tunnel_testnet_community():
-    from tribler_core.modules.tunnel.community.community import TriblerTunnelTestnetCommunity
-    return TriblerTunnelTestnetCommunity
-
-
 # strategies
 def random_churn():
     from ipv8.peerdiscovery.churn import RandomChurn
@@ -99,16 +64,6 @@ def periodic_similarity():
     return PeriodicSimilarity
 
 
-def golden_ratio_strategy():
-    from tribler_core.modules.tunnel.community.discovery import GoldenRatioStrategy
-    return GoldenRatioStrategy
-
-
-def remove_peers():
-    from tribler_core.modules.metadata_store.community.sync_strategy import RemovePeers
-    return RemovePeers
-
-
 @precondition('session.config.discovery_community.enabled')
 @overlay(discovery_community)
 @kwargs(max_peers='100')
@@ -119,24 +74,6 @@ class IPv8DiscoveryCommunityLauncher(IPv8CommunityLauncher):
     pass
 
 
-@overlay(bandwidth_accounting_community)
-@precondition('not session.bandwidth_testnet()')
-@walk_strategy(random_walk)
-@set_in_session('bandwidth_community')
-class BandwidthCommunityLauncher(IPv8CommunityLauncher):
-    def get_kwargs(self, session):
-        return {
-            'settings': session.config.bandwidth_accounting,
-            'database_path': session.config.state_dir / "sqlite" / "bandwidth.db",
-        }
-
-
-@overlay(bandwidth_accounting_testnet_community)
-@precondition('session.bandwidth_testnet()')
-class BandwidthTestnetCommunityLauncher(TestnetMixIn, BandwidthCommunityLauncher):
-    pass
-
-
 @precondition('session.config.dht.enabled')
 @set_in_session('dht_community')
 @overlay(dht_discovery_community)
@@ -144,83 +81,6 @@ class BandwidthTestnetCommunityLauncher(TestnetMixIn, BandwidthCommunityLauncher
 @walk_strategy(ping_churn, target_peers=INFINITE)
 @walk_strategy(random_walk)
 class DHTCommunityLauncher(IPv8CommunityLauncher):
-    pass
-
-
-@after('DHTCommunityLauncher', 'BandwidthCommunityLauncher', 'BandwidthTestnetCommunityLauncher')
-@precondition('session.config.tunnel_community.enabled')
-@precondition('not session.tunnel_testnet()')
-@set_in_session('tunnel_community')
-@overlay(tribler_tunnel_community)
-@walk_strategy(random_walk)
-@walk_strategy(golden_ratio_strategy, target_peers=INFINITE)
-class TriblerTunnelCommunityLauncher(IPv8CommunityLauncher):
-    def get_kwargs(self, session):
-        from ipv8.dht.provider import DHTCommunityProvider
-        from ipv8.messaging.anonymization.community import TunnelSettings
-
-        settings = TunnelSettings()
-        settings.min_circuits = session.config.tunnel_community.min_circuits
-        settings.max_circuits = session.config.tunnel_community.max_circuits
-
-        return {
-            'bandwidth_community': session.bandwidth_community,
-            'competing_slots': session.config.tunnel_community.competing_slots,
-            'ipv8': session.ipv8,
-            'random_slots': session.config.tunnel_community.random_slots,
-            'tribler_session': session,
-            'dht_provider': DHTCommunityProvider(session.dht_community, session.config.ipv8.port),
-            'settings': settings
-        }
-
-
-@precondition('session.config.tunnel_community.enabled')
-@precondition('session.tunnel_testnet()')
-@overlay(tribler_tunnel_testnet_community)
-class TriblerTunnelTestnetCommunityLauncher(TestnetMixIn, TriblerTunnelCommunityLauncher):
-    pass
-
-
-@precondition('session.config.popularity_community.enabled')
-@set_in_session('popularity_community')
-@overlay(popularity_community)
-@walk_strategy(random_walk, target_peers=30)
-@walk_strategy(remove_peers, target_peers=INFINITE)
-class PopularityCommunityLauncher(IPv8CommunityLauncher):
-    def get_kwargs(self, session):
-        return {
-            'settings': session.config.popularity_community,
-            'rqc_settings': session.config.remote_query_community,
-
-            'metadata_store': session.mds,
-            'torrent_checker': session.torrent_checker,
-        }
-
-
-@overlay(giga_channel_community)
-@precondition('session.config.chant.enabled')
-@precondition('not session.chant_testnet()')
-@set_in_session('gigachannel_community')
-@overlay(giga_channel_community)
-# GigaChannelCommunity remote search feature works better with higher amount of connected peers
-@walk_strategy(random_walk, target_peers=30)
-@walk_strategy(remove_peers, target_peers=INFINITE)
-class GigaChannelCommunityLauncher(IPv8CommunityLauncher):
-    def get_kwargs(self, session):
-        return {
-            'settings': session.config.chant,
-            'rqc_settings': session.config.remote_query_community,
-
-            'metadata_store': session.mds,
-            'notifier': session.notifier,
-            'max_peers': 50,
-        }
-
-
-@precondition('session.config.chant.enabled')
-@precondition('session.chant_testnet()')
-@overlay(giga_channel_testnet_community)
-class GigaChannelTestnetCommunityLauncher(TestnetMixIn, GigaChannelCommunityLauncher):
     pass
 
 
@@ -246,11 +106,25 @@ def register_default_launchers(loader):
      the CommunityLaunchers themselves to be properly scoped and lazy-loaded.
     """
     loader.set_launcher(IPv8DiscoveryCommunityLauncher())
-    loader.set_launcher(BandwidthCommunityLauncher())
-    loader.set_launcher(BandwidthTestnetCommunityLauncher())
     loader.set_launcher(DHTCommunityLauncher())
+
+    from tribler_core.modules.bandwidth_accounting.launcher import BandwidthCommunityLauncher
+    loader.set_launcher(BandwidthCommunityLauncher())
+
+    from tribler_core.modules.bandwidth_accounting.launcher import BandwidthTestnetCommunityLauncher
+    loader.set_launcher(BandwidthTestnetCommunityLauncher())
+
+    from tribler_core.modules.tunnel.community.launcher import TriblerTunnelCommunityLauncher
     loader.set_launcher(TriblerTunnelCommunityLauncher())
+
+    from tribler_core.modules.tunnel.community.launcher import TriblerTunnelTestnetCommunityLauncher
     loader.set_launcher(TriblerTunnelTestnetCommunityLauncher())
+
+    from tribler_core.modules.popularity.launcher import PopularityCommunityLauncher
     loader.set_launcher(PopularityCommunityLauncher())
+
+    from tribler_core.modules.metadata_store.community.launcher import GigaChannelCommunityLauncher
     loader.set_launcher(GigaChannelCommunityLauncher())
+
+    from tribler_core.modules.metadata_store.community.launcher import GigaChannelTestnetCommunityLauncher
     loader.set_launcher(GigaChannelTestnetCommunityLauncher())
